Fix build issues and update target to CommonJS (#21)
* fix build errors

* fix tsconfig.json options

* update examples

* add back declarations

+++ packages/react-app-server/build.ts
@@ -113,6 +113,10 @@ async function build(
 }
 
 function copyPublicFolder() {
+  if (!fs.existsSync(paths.appPublic)) {
+    return
+  }
+
   fs.copySync(paths.appPublic, paths.appDistPublic, {
     dereference: true,
   })
